갤러리 선택 주제 보이기, 관리페이지 권한 수정

+++ back/src/controller/adminController.js
@@ -4,13 +4,17 @@ import Post from '../model/Post';
 //포스트와 회원목록 불러오기
 export const loadPostAndUser = async (req, res) => {
   try {
-    const post = await Post.find({}).populate('creator').sort({ _id: -1 });
-    const user = await User.find({}).sort({ _id: -1 });
-    const allList = {
-      post,
-      user,
-    };
-    res.json(allList);
+    if (req.user.role === 'admin') {
+      const post = await Post.find({}).populate('creator').sort({ _id: -1 });
+      const user = await User.find({}).sort({ _id: -1 });
+      const allList = {
+        post,
+        user,
+      };
+      return res.json(allList);
+    } else {
+      return res.status(404).send('권한 없음');
+    }
   } catch (e) {
     console.log(e);
     res.status(400);
@@ -21,9 +25,13 @@ export const loadPostAndUser = async (req, res) => {
 export const secession = async (req, res) => {
   const { id } = req.body;
   try {
-    await User.findByIdAndRemove(id);
-    const userList = await User.find({}).sort({ _id: -1 });
-    res.json(userList);
+    if (req.user.role === 'admin') {
+      await User.findByIdAndRemove(id);
+      const userList = await User.find({}).sort({ _id: -1 });
+      return res.json(userList);
+    } else {
+      return res.status(404).send('권한 없음');
+    }
   } catch (e) {
     console.log(e);
     res.status(400);
@@ -34,14 +42,37 @@ export const secession = async (req, res) => {
 export const editRole = async (req, res) => {
   const { objectId, role } = req.body;
   try {
-    console.log(objectId, role);
-    await User.findByIdAndUpdate(objectId, {
-      role,
-    });
-    const newUserList = await User.find({}).sort({ _id: -1 })
-    return res.json(newUserList);
+    if (req.user.role === 'admin') {
+      await User.findByIdAndUpdate(objectId, {
+        role,
+      });
+      const newUserList = await User.find({}).sort({ _id: -1 });
+      return res.json(newUserList);
+    } else {
+      return res.status(404).send('권한 없음');
+    }
   } catch (e) {
     console.log(e);
     res.status(400);
   }
+};
+
+//관리페이지에서 포스트 삭제
+
+export const deletePostFromAdmin = async (req, res) => {
+  const { id } = req.body;
+  try {
+    if (req.user.role === 'admin') {
+      await Post.findByIdAndRemove({ _id: id });
+      const PostList = await Post.find({})
+        .populate('creator')
+        .sort({ _id: -1 });
+      return res.json(PostList);
+    } else {
+      return res.status(404).send('권한 없음');
+    }
+  } catch (e) {
+    console.log(e);
+  }
+  return;
 };
